+++ scripts/vfx/build_rawtoaces.sh
@@ -20,11 +20,35 @@ fi
 tar -zxf "$DOWNLOADS_DIR/rawtoaces-${ASWF_RAWTOACES_VERSION_DOWNLOAD}.tar.gz"
 cd "rawtoaces-${ASWF_RAWTOACES_VERSION_DOWNLOAD}"
 
+if [[ $ASWF_RAWTOACES_VERSION == 2.1.1 ]]; then
+
+# Don't use single quotes around EOF to allow ASWF_CONAN_PYTHON_VERSION to be expanded
+cat << 'EOF' | patch -p1
+diff --git a/configure.cmake b/configure.cmake
+index d99dd79ac..bfda2778a 100644
+--- a/configure.cmake
++++ b/configure.cmake
+@@ -25,7 +25,7 @@
+         set( DEV_MODULE Development.Module )
+     endif ()
+ 
+-    find_package ( Python 3.8 COMPONENTS Interpreter ${DEV_MODULE} REQUIRED )
++    find_package ( Python ${RTA_PYTHON_VERSION} EXACT COMPONENTS Interpreter ${DEV_MODULE} REQUIRED )
+ 
+     execute_process (
+         COMMAND "\${Python_EXECUTABLE}" -m nanobind --cmake_dir
+EOF
+
+fi
+
 mkdir build
 cd build
 cmake \
      -DCMAKE_INSTALL_PREFIX="${ASWF_INSTALL_PREFIX}" \
-     -DRTA_BUILD_PYTHON_BINDINGS=FALSE \
+     -DRTA_BUILD_PYTHON_BINDINGS="ON" \
+     -DRTA_PYTHON_VERSION="${ASWF_CONAN_PYTHON_VERSION}" \
+     -DRTA_ENABLE_LENSFUN="ON" \
+     -DRTA_ENABLE_EIGEN="ON" \
      ..
 cmake --build . -j$(nproc)
 cmake --install .